Move messages into Rooms
